[Refactor] Relocate to `domain.place.cluster.vo`
