+++ src/main/java/boombimapi/domain/place/cluster/vo/ClusterInput.java
@@ -1,4 +1,4 @@
-package boombimapi.global.geo.core;
+package boombimapi.domain.place.cluster.vo;
 
 public record ClusterInput(
     Long id,
